Add support for more compile filters
Before it was possible to control the build only with `build_lib` and
`build_bin` which allowed for only one thing to be built at a time. This
commit adds support for other types of builds (such as examples and
tests) by names (as if running `cargo check --example some_example`) as
well as building all targets of a certain type (as if running `cargo check
--examples`) and all targets (as if running `cargo check --all-targets`).

Backward compatibility with clients using `build_lib` and `build_bin` is
preserved.